Cant hold it anymore

+++ src/shared/infra/typeorm/migrations/1632249171262-CreateBundleRelations.ts
@@ -27,7 +27,7 @@ export class CreateBundleRelations1632249171262 implements MigrationInterface {
 					},
 					{
 						name: 'percentage',
-						type: 'int',
+						type: 'decimal',
 						isNullable: false,
 						isUnique: false,
 					},
